1.6.43

+++ AffiseAttributionLib/Classes/usecase/SendDataToServerUseCaseImpl.swift
@@ -13,14 +13,18 @@ internal class SendDataToServerUseCaseImpl {
 //    private let metricsRepository: MetricsRepository
     private let logsManager: LogsManager
     private let preferencesUseCase: PreferencesUseCase
+    private let firstAppOpenUseCase: FirstAppOpenUseCase
     
-    init(postBackModelFactory: PostBackModelFactory,
-         cloudRepository: CloudRepository,
-         eventsRepository: EventsRepository,
-         internalEventsRepository: InternalEventsRepository,
-         logsRepository: LogsRepository,
-         logsManager: LogsManager,
-         preferencesUseCase: PreferencesUseCase) {
+    init(
+        postBackModelFactory: PostBackModelFactory,
+        cloudRepository: CloudRepository,
+        eventsRepository: EventsRepository,
+        internalEventsRepository: InternalEventsRepository,
+        logsRepository: LogsRepository,
+        logsManager: LogsManager,
+        preferencesUseCase: PreferencesUseCase,
+        firstAppOpenUseCase: FirstAppOpenUseCase
+    ) {
         
         self.postBackModelFactory = postBackModelFactory
         self.cloudRepository = cloudRepository
@@ -29,6 +33,7 @@ internal class SendDataToServerUseCaseImpl {
         self.logsRepository = logsRepository
         self.logsManager = logsManager
         self.preferencesUseCase = preferencesUseCase
+        self.firstAppOpenUseCase = firstAppOpenUseCase
     }
 
     /**
@@ -53,7 +58,7 @@ internal class SendDataToServerUseCaseImpl {
     /**
      * Sending for url
      */
-    private func send(url: String) {
+    private func send(url: String, sendEmpty: Bool) {
         
         var events: Array<SerializedEvent> = []
         var logs: Array<SerializedLog> = []
@@ -73,13 +78,29 @@ internal class SendDataToServerUseCaseImpl {
             //Get internal events
             internalEvents = internalEventsRepository.getEvents(url: url)
 
+            if !sendEmpty && !(!events.isEmpty || !logs.isEmpty || !internalEvents.isEmpty) {
+                // if flag sendEmpty is false and all array is empty
+                // don't send empty postback
+                return
+            }
+
             //Generate data
-            let data = [postBackModelFactory.create(events: events, logs: logs, metrics: metrics, internalEvents: internalEvents)]
+            var data = postBackModelFactory.create(events: events, logs: logs, metrics: metrics, internalEvents: internalEvents)
             
+            // If first run
+            if (firstAppOpenUseCase.isFirstOpen()) {
+                data = data.asFirstOpen()
+            }
+
             do {
                 //Send data for single url
-                try cloudRepository.send(data: data, url: url)
+                try cloudRepository.send(data: [data], url: url)
                 
+                if (firstAppOpenUseCase.isFirstOpen()) {
+                    // Complete first open
+                    firstAppOpenUseCase.completeFirstOpen()
+                }
+
                 //Remove sent events
                 eventsRepository.deleteEvent(ids: events.map { it in it.id }, url: url)
                 
@@ -105,7 +126,7 @@ extension SendDataToServerUseCaseImpl: SendDataToServerUseCase {
     /**
      * Send data
      */
-    func send(withDelay: Bool) {
+    func send(withDelay: Bool, sendEmpty: Bool) {
         if (preferencesUseCase.isOfflineModeEnabled()) {
             logsManager.addSdkError(error: AffiseError.offlineModeEnabled)
             return
@@ -124,7 +145,7 @@ extension SendDataToServerUseCaseImpl: SendDataToServerUseCase {
                         Thread.sleep(forTimeInterval: SendDataToServerUseCaseImpl.TIME_DELAY_SENDING)
                     }
 
-                    self?.send(url: it)
+                    self?.send(url: it, sendEmpty: sendEmpty)
 
                     //Unlock sending to this url
                     self?.setIsSend(it, false)
